+++ backend/src/main/java/org/vaccom/vcmgt/action/CongDanAction.java
@@ -1,0 +1,9 @@
+package org.vaccom.vcmgt.action;
+
+import org.springframework.stereotype.Service;
+import org.vaccom.vcmgt.entity.CongDan;
+@Service
+public interface CongDanAction {
+    public CongDan findByCongDanId(long id);
+    public CongDan update(CongDan congDan);
+}