+++ backend/src/main/java/org/vaccom/vcmgt/action/NguoiTiemChungAction.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.vaccom.vcmgt.dto.NguoiTiemChungDto;
 import org.vaccom.vcmgt.dto.ResultSearchDto;
+import org.vaccom.vcmgt.entity.CongDan;
 import org.vaccom.vcmgt.entity.NguoiTiemChung;
 
 @Service
@@ -53,5 +54,11 @@ public interface NguoiTiemChungAction {
 
 	public ResultSearchDto<NguoiTiemChung> search(NguoiTiemChungDto nguoiTiemChungDto, int page, int size);
 
+	public ResultSearchDto<CongDan> searchDaTiem(NguoiTiemChungDto nguoiTiemChungDto, int page, int size);
+
 	public NguoiTiemChung findByCMTCCCD(String cmtcccd);
+
+	public NguoiTiemChung update(NguoiTiemChung nguoiTiemChung);
+
+	public NguoiTiemChung findBycongDanID(long id);
 }