+++ backend/src/main/java/org/vaccom/vcmgt/entity/MuiTiemChung.java
@@ -28,9 +28,6 @@ public class MuiTiemChung implements Serializable {
 	@Column(name = "ID")
 	private long id;
 
-	@Column(name = "NguoiTiemChungID")
-	private long nguoiTiemChungId;
-
 	@Column(name = "CoSoYTeID")
 	private long coSoYTeId;
 
@@ -72,6 +69,17 @@ public class MuiTiemChung implements Serializable {
 
 	@Column(name = "HanSuDung")
 	private String hanSuDung;
+
+	@Column(name = "CongDan_ID")
+	private Long congDanID;
+
+	public Long getCongDanID() {
+		return congDanID;
+	}
+
+	public void setCongDanID(Long congDanID) {
+		this.congDanID = congDanID;
+	}
 
 	public long getId() {
 		return id;
@@ -183,14 +191,6 @@ public class MuiTiemChung implements Serializable {
 
 	public void setHanSuDung(String hanSuDung) {
 		this.hanSuDung = hanSuDung;
-	}
-
-	public long getNguoiTiemChungId() {
-		return nguoiTiemChungId;
-	}
-
-	public void setNguoiTiemChungId(long nguoiTiemChungId) {
-		this.nguoiTiemChungId = nguoiTiemChungId;
 	}
 
 	public long getCoSoYTeId() {