+++ backend/src/main/java/org/vaccom/vcmgt/service/MuiTiemChungService.java
@@ -15,7 +15,7 @@ public interface MuiTiemChungService {
 	
 	public long countByCoSoYTeId(long id);
 	
-	public long countByNguoiTiemChungId(long id);
+	public long countByCongDan_ID(long id);
 	
 	public void deleteById(long id);
 	
@@ -24,8 +24,8 @@ public interface MuiTiemChungService {
 	public MuiTiemChung findById(long id);
 	
 	public List<MuiTiemChung> findByCmtcccd(String cmtcccd);
-	
-	public List<MuiTiemChung> findByNguoiTiemChungId(long id);
+
+	public List<MuiTiemChung> findByCongDan_ID(long id);
 
 	public List<MuiTiemChung> findByCoSoYTeId(long id, int page, int size);