+++ backend/src/main/resources/schema-mysql.sql
@@ -15,6 +15,7 @@ CREATE TABLE IF NOT EXISTS `t_nguoidung` (
   `NguoiTiemChungID` bigint(20) DEFAULT 0,
   `UyBanNhanDanID` bigint(20) DEFAULT 0,
   `QuanTriHeThong` tinyint(4) DEFAULT 0,
+  `CongDan_ID`  bigint(20) DEFAULT 0,
   PRIMARY KEY (`ID`)
 ) ;
 
@@ -130,9 +131,9 @@ CREATE TABLE IF NOT EXISTS `t_muitiemchung` (
   `SoLoThuoc` varchar(64) DEFAULT NULL,
   `HanSuDung` varchar(10) DEFAULT NULL,
   `CoSoYTeID` bigint(20) DEFAULT 0,
-  `NguoiTiemChungID` bigint(20) NOT NULL,
+  `CongDan_ID`  bigint(20) DEFAULT 0,
   PRIMARY KEY (`ID`)
-) ;
+);
 
 CREATE TABLE IF NOT EXISTS `t_nguoitiemchung` (
   `ID` bigint(20) NOT NULL AUTO_INCREMENT,
@@ -167,6 +168,9 @@ CREATE TABLE IF NOT EXISTS `t_nguoitiemchung` (
   `MaQR` varchar(128) NOT NULL,
   `KiemTraTrung` tinyint(4) DEFAULT 0,
   `KetQuaKiemTra` longtext DEFAULT NULL,
+  `CongDan_ID`  bigint(20) DEFAULT 0,
+  `SoMuiTiem` tinyint(4) DEFAULT 0,
+  `NgayTiemCuoi` varchar(20) DEFAULT NULL,
   PRIMARY KEY (`ID`)
 ) ;
 
@@ -199,8 +203,11 @@ CREATE TABLE IF NOT EXISTS `t_giaydiduong` (
   `CheckSum` text DEFAULT NULL,
   `GhiChu` text DEFAULT NULL,
   `Status` tinyint(1) DEFAULT 0,
+  `CongDan_ID` bigint(20) DEFAULT 0,
   PRIMARY KEY (`ID`)
 ) ;
+
+
 
 CREATE TABLE IF NOT EXISTS `t_thuoc` (
   `ID` bigint(20) NOT NULL AUTO_INCREMENT,
@@ -301,4 +308,26 @@ create table IF NOT EXISTS t_hangchothongbao (
     `createDate` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     primary key (ID)
 );
+
+create table IF NOT EXISTS t_congdan (
+    `ID` bigint not null auto_increment,
+    `HoVaTen` varchar(128) NOT NULL,
+    `NgaySinh` varchar(10) DEFAULT NULL,
+    `GioiTinh` tinyint(4) DEFAULT 0,
+    `DiaChiThuongTru` varchar(512) DEFAULT NULL,
+    `TinhThanh_Ma` varchar(30) DEFAULT NULL,
+    `TinhThanh_Ten` varchar(128) DEFAULT NULL,
+    `QuanHuyen_Ma` varchar(30) DEFAULT NULL,
+    `QuanHuyen_Ten` varchar(128) DEFAULT NULL,
+    `PhuongXa_Ma` varchar(30) DEFAULT NULL,
+    `PhuongXa_Ten` varchar(128) DEFAULT NULL,
+    `CMTCCCD` varchar(20) DEFAULT NULL,
+    `NoiCap` varchar(128) DEFAULT NULL,
+    `NgayCap` varchar(20) DEFAULT NULL,
+    `SoMuiTiem` tinyint(4) DEFAULT 0,
+    `SoDienThoai` varchar(20) DEFAULT NULL,
+    primary key (ID)
+)
+
+